+++ src/svc/prg-ustat.cpp
@@ -37,7 +37,7 @@ static std::string exec(const char* cmd) {
     FILE* pipe = popen(cmd, "r");
     if (!pipe) throw std::runtime_error("popen() failed!");
     try {
-        while (fgets(buffer, sizeof buffer, pipe) != NULL) {
+        while (fgets(buffer, sizeof buffer, pipe) != nullptr) {
             result += buffer;
         }
     } catch (...) {