+++ src/svc/svc.cpp
@@ -33,6 +33,7 @@ ThreadSafeQueue<string>                 log_util = ThreadSafeQueue<string>();
 unordered_map<string, UnifiedService>   services;
 
 // PRIVATE VARIABLES
+unordered_map<string, future<int>>      futures;
 unordered_map<string, future<int>>      system_services_futures;
 unordered_map<string, UnifiedService>   system_services;
 
@@ -66,11 +67,9 @@ bool addSystemRoutine(const string& id, const function<int(const vector<string>&
 }
 
 bool requestRoutineStatus(const string& id) {
-    if (services.contains(id)) {
+    if (services.contains(id))
         return services.at(id).getStatus();
-    } else {
-        return true;
-    }
+    return true;
 }
 
 function<int(const vector<string>&)> requestRoutine(const string& id) {
@@ -81,9 +80,13 @@ bool requestValidRoutineID(const string& id) {
     return services.contains(id);
 }
 
-bool requestRunRoutine(const string& id) {
+bool requestRunRoutine(const vector<string>& cmd) {
+    string id = cmd.at(1);
     if (services.contains(id)) {
-        return services.at(id).run();
+        services.at(id).run();
+        futures.emplace(id, std::async(std::launch::async, [cmd, id] { return requestRoutine(
+                            id)(cmd); }));
+        return true;
     } else {
         return false;
     }
@@ -92,19 +95,26 @@ bool requestRunRoutine(const string& id) {
 bool requestStopRoutine(const string& id) {
     if (services.contains(id)) {
         services.at(id).stop();
+        if (futures.contains(id))
+            futures.erase(id);
+        log_util.push("#O [0] <svc.cpp>: unordered_map<string, future<int>> futures length: " + std::to_string(futures.size()));
         return true;
     } else {
         return false;
     }
 }
 
-// This function will be run in monitor_futures() thread in boot.cpp
+void show_active_services() {
+    log_util.push("#S [0] <svc.cpp>: Active Services: " + std::to_string(futures.size()));
+}
+
+// This function will be run in monitor_futures() thread
 // It does not have its own thread in order to save resources
 void monitor_system_futures() {
     for (auto it = system_services_futures.begin(); it != system_services_futures.end(); ) {
         if (it->second.wait_for(std::chrono::seconds(0)) == std::future_status::ready) {
             string service_id = it->first;
-            requestStopRoutine(service_id);
+            system_services.at(service_id).stop();
             it = system_services_futures.erase(it);
             if (service_id == "security.svc") {
                 log_system.push("#Security [monitor_system_futures()] <svc.cpp>: security.svc has been removed from system_services_futures! This is usually a result of an early return due to an error. Please restart PXL.");
@@ -117,6 +127,35 @@ void monitor_system_futures() {
             ++it;
         }
     }
+}
+
+// This function will be run in a separate thread
+void monitor_futures() {
+    log_system.push("#O [monitor_futures()] <svc.cpp>: Futures are now being monitored every 0.25 seconds in a separate thread");
+    while (true) {
+        // Monitor system futures of svc.cpp
+        monitor_system_futures();
+
+        for (auto it = futures.begin(); it != futures.end(); ) {
+            if (it->second.wait_for(std::chrono::seconds(0)) == std::future_status::ready) {
+                string service_id = it->first;
+                services.at(service_id).stop();
+                it = futures.erase(it);
+                log_system.push("#O [monitor_futures()] <svc.cpp>: [" + service_id + "] has been erased from futures");
+                log_system.push("#O [monitor_futures()] <svc.cpp>: Futures Length: " + std::to_string(futures.size()));
+            } else {
+                ++it;
+            }
+        }
+
+        // Sleep for a reasonable interval before checking again
+        std::this_thread::sleep_for(std::chrono::milliseconds(250));
+    }
+}
+
+void start_monitor_futures() {
+    std::thread monitorThread(monitor_futures);
+    monitorThread.detach(); // detach so it doesn't result in terminate called without an active exception
 }
 
 void makeRoutines() {